Update patch set 1
Patch Set 1:

swagger.json does not change once release, code is going to generate it every time the same json. to avoid this, swagger-sdk bring the cocept that generate once as part of build and use it ever. because ci does not have ability to check-in this json, we the community should put into the code base.

Patch-set: 1